Rearrange code and make good structure, Fix Banner Ads size and position
