+++ app/src/main/java/wallpaper/deekshithrajbasa/com/wall_papers/category/material.java
@@ -1,4 +1,4 @@
-package wallpaper.deekshithrajbasa.com.wall_papers;
+package wallpaper.deekshithrajbasa.com.wall_papers.category;
 
 import android.content.Intent;
 import android.graphics.Bitmap;
@@ -21,16 +21,20 @@ import android.widget.TextView;
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.squareup.picasso.Picasso;
 
+import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.ArrayList;
 
 import jp.wasabeef.recyclerview.animators.SlideInUpAnimator;
+import wallpaper.deekshithrajbasa.com.wall_papers.R;
+import wallpaper.deekshithrajbasa.com.wall_papers.utils.SimpleDividerItemDecoration;
+import wallpaper.deekshithrajbasa.com.wall_papers.utils.customview;
+import wallpaper.deekshithrajbasa.com.wall_papers.adapter.instagramAdapter;
 
-public class flower extends AppCompatActivity {
+public class material extends AppCompatActivity {
     public static final String EXTRA_URL = "imageurl";
     DatabaseReference dref;
     ListView listview;
@@ -47,13 +51,15 @@ public class flower extends AppCompatActivity {
     private ImageButton mSearchBtn;
     //arrays to store image, title, description, int position(0,1,2,3..)
     public static ArrayList<String> imageUrl = new ArrayList<>();
+
     public static ArrayList<Integer> pos = new ArrayList<>();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_flower);
+        setContentView(R.layout.activity_material);
         Thread.setDefaultUncaughtExceptionHandler(new material.MyUncaughtExceptionHandler());
+
 
         recyclerView = (RecyclerView) findViewById(R.id.Recycleview);
         recyclerView.setHasFixedSize(true);
@@ -64,20 +70,20 @@ public class flower extends AppCompatActivity {
         gridLayoutManager.setOrientation(LinearLayoutManager.VERTICAL); // set Horizontal Orientation
         recyclerView.setLayoutManager(gridLayoutManager);
         //firebase
-        myref = FirebaseDatabase.getInstance().getReference().child("/flower");
-        FirebaseRecyclerAdapter<instagramAdapter,flower.BlogViewHolder> recyclerAdapter = new FirebaseRecyclerAdapter<instagramAdapter, flower.BlogViewHolder>(
+        myref = FirebaseDatabase.getInstance().getReference().child("/material");
+        FirebaseRecyclerAdapter<instagramAdapter,material.BlogViewHolder> recyclerAdapter = new FirebaseRecyclerAdapter<instagramAdapter, material.BlogViewHolder>(
                 instagramAdapter.class,
                 R.layout.individual_row,
-                flower.BlogViewHolder.class,
+                material.BlogViewHolder.class,
                 myref
+
         )
         {
             @Override
-            protected void populateViewHolder(flower.BlogViewHolder viewHolder, instagramAdapter model, int position) {
+            protected void populateViewHolder(material.BlogViewHolder viewHolder, instagramAdapter model, int position) {
 
                 //Lv-Edited
                 viewHolder.imageView.setImageDrawable(getResources().getDrawable(R.drawable.loadingpic));
-
                 viewHolder.setImage(model.getImage());
                 viewHolder.setPosition(position);
                 imageUrl.add("" + model.getImage());
@@ -86,12 +92,28 @@ public class flower extends AppCompatActivity {
                 if(!pos.contains(position))
                     pos.add(position);
 
+
             }
         };
         recyclerView.setAdapter(recyclerAdapter);
 
 
 
+    }
+    public static class MyUncaughtExceptionHandler implements Thread.UncaughtExceptionHandler {
+        @Override
+        public void uncaughtException(Thread thread, Throwable ex) {
+            if(ex.getClass().equals(OutOfMemoryError.class))
+            {
+                try {
+                    android.os.Debug.dumpHprofData("/sdcard/dump.hprof");
+                }
+                catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+            ex.printStackTrace();
+        }
     }
 
     public static class BlogViewHolder extends RecyclerView.ViewHolder {
@@ -183,6 +205,7 @@ public class flower extends AppCompatActivity {
             }
 
         }
+
     }
     }
