+++ app/src/main/java/wallpaper/deekshithrajbasa/com/wall_papers/utils/BuildingFragment.java
@@ -1,4 +1,4 @@
-package wallpaper.deekshithrajbasa.com.wall_papers;
+package wallpaper.deekshithrajbasa.com.wall_papers.utils;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -7,6 +7,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import wallpaper.deekshithrajbasa.com.wall_papers.R;
 
 public class BuildingFragment extends Fragment {
