+++ app/src/main/java/wallpaper/deekshithrajbasa/com/wall_papers/utils/SimpleDividerItemDecoration.java
@@ -1,10 +1,12 @@
-package wallpaper.deekshithrajbasa.com.wall_papers;
+package wallpaper.deekshithrajbasa.com.wall_papers.utils;
 
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.drawable.Drawable;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import wallpaper.deekshithrajbasa.com.wall_papers.R;
 
 public class SimpleDividerItemDecoration extends RecyclerView.ItemDecoration {
     private Drawable mDivider;
