+++ app/src/main/res/layout/activity_building.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".building">
+    tools:context=".category.building">
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app"
         android:layout_width="match_parent"
