+++ app/src/main/res/layout/activity_customview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".customview">
+    tools:context=".utils.customview">
 
     <ImageView
         android:id="@+id/ImageView"
