+++ app/src/main/res/layout/activity_flower.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".flower">
+    tools:context=".category.flower">
     <android.support.v7.widget.RecyclerView
         android:id="@+id/Recycleview"
         android:layout_width="match_parent"
