+++ app/src/main/res/layout/app_bar_main.xml
@@ -21,8 +21,6 @@
 
         </android.support.v7.widget.Toolbar>
 
-
-
     </android.support.design.widget.AppBarLayout>
 
     <include layout="@layout/content_main" />
